move Common.Extensions To Common.Utils
