+++ SCHALE.Toolbox/Forms/PlayableCharacterForm.cs
@@ -9,8 +9,8 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using SCHALE.Common.Extensions;
 using SCHALE.Common.FlatData;
+using SCHALE.Common.Utils;
 using SCHALE.GameServer.Services;
 using SCHALE.Toolbox.Models.SchaleDB;
 using Serilog;
